services/finops-dashboard/requirements.txt
# FinOps Dashboard - Python Dependencies

# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# AWS integration
boto3==1.34.34
botocore==1.34.34

# GCP integration
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-auth==2.27.0

# Azure integration (future)
# azure-mgmt-costmanagement==4.0.1
# azure-identity==1.15.0

# Data processing
pandas==2.2.0
numpy==1.26.3

# ML & Forecasting
prophet==1.1.5
scikit-learn==1.4.0
scipy==1.12.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==24.1.1
ruff==0.2.0
mypy==1.8.0

# Auth0 / JWT
PyJWT[crypto]==2.8.0
python-jose[cryptography]==3.3.0
cryptography==42.0.2

# Alembic (Database migrations)
alembic==1.13.1

# Azure SDK
azure-identity==1.15.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-monitor==6.0.2
azure-core==1.30.0
